Don't conflict with stdlib secrets module Setup Click.
