+++ .gitignore
@@ -1,0 +1,4 @@
+*.egg-info
+**/__pycache__
+
+.env
